GUI: fix 3rd float accuracy issue
[JIRA] BTSOCGUI-634

+++ example/application/screen_410_502/app_3d_pondKoi.c
@@ -472,7 +472,7 @@ static void wave_animate_cb(void *parent)
             wave_animating = true;
         }
 
-        gui_log("Wave center: (%f, %f)\n", wave_x, wave_y);
+        gui_log("Wave center: (%f, %f)\n", (double)wave_x, (double)wave_y);
         // matrix_translate(wave_x, wave_x, wave_win->base.matrix);
 
     }
